requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.1
django-admin-interface==0.29.4
django-admin-list-filter-dropdown==1.0.3
django-admin-numeric-filter==0.1.9
django-admin-rangefilter==0.13.2
django-colorfield==0.12.0
django-cors-headers==4.4.0
django-filter==25.1
django-jazzmin==2.6.0
django-jet==1.0.8
django-jet-reboot==1.3.10
django-material==1.12.0
django-material-admin==1.8.6
django-viewflow==2.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
frozenlist==1.4.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
magic-filter==1.0.12
multidict==6.1.0
pillow==11.1.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.31.0
rpds-py==0.20.0
six==1.17.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.11.1
